added cy.arrow command
